src/python/requirements.txt
torch>=2.2
numpy>=1.24
Pillow>=10.0
cairosvg>=2.7
